Move kernel/execution code out of NotebookEditorWidget for testability
